// src/youtube/types/ArtistResults_0.ts
export type ArtistResults_0 = Root2[]

export interface Root2 {
    responseContext: ResponseContext
    items?: Item[]
    trackingParams: string
    contents: Contents
    header?: Header3
}

export interface ResponseContext {
    serviceTrackingParams: ServiceTrackingParam[]
    maxAgeSeconds?: number
}

export interface ServiceTrackingParam {
    service: string
    params: Param[]
}

export interface Param {
    key: string
    value: string
}

export interface Item {
    guideSectionRenderer: GuideSectionRenderer
}
// ...
export interface Contents {
    singleColumnBrowseResultsRenderer: SingleColumnBrowseResultsRenderer
}
// ...
export interface MusicShelfRenderer {
    title: Title
    contents: Content3[]
    trackingParams: string
    bottomText: BottomText
    bottomEndpoint: BottomEndpoint
    shelfDivider: ShelfDivider
}

export interface Title {
    runs: Run8[]
}

export interface Run8 {
    text: string
    navigationEndpoint: NavigationEndpoint3
}
// ...
export interface Content3 {
    musicResponsiveListItemRenderer: MusicResponsiveListItemRenderer
}

export interface MusicResponsiveListItemRenderer {
    trackingParams: string
    thumbnail: Thumbnail
    flexColumns: FlexColumn[]
    menu: Menu
    playlistItemData: PlaylistItemData
    flexColumnDisplayStyle: string
    navigationEndpoint: NavigationEndpoint6
}
// ...
export interface BottomText {
    runs: Run15[]
}

export interface Run15 {
    text: string
}

export interface BottomEndpoint {
    clickTrackingParams: string
    browseEndpoint: BrowseEndpoint4
}

export interface BrowseEndpoint4 {
    browseId: string
    params: string
    browseEndpointContextSupportedConfigs: BrowseEndpointContextSupportedConfigs4
}

export interface BrowseEndpointContextSupportedConfigs4 {
    browseEndpointContextMusicConfig: BrowseEndpointContextMusicConfig4
}

export interface BrowseEndpointContextMusicConfig4 {
    pageType: string
}

export interface ShelfDivider {
    musicShelfDividerRenderer: MusicShelfDividerRenderer
}

export interface MusicShelfDividerRenderer {
    hidden: boolean
}

export interface MusicCarouselShelfRenderer {
    header: Header2
    contents: Content4[]
    trackingParams: string
    itemSize: string
}

export interface Header2 {
    musicCarouselShelfBasicHeaderRenderer: MusicCarouselShelfBasicHeaderRenderer
}

export interface MusicCarouselShelfBasicHeaderRenderer {
    title: Title2
    accessibilityData: AccessibilityData5
    headerStyle: string
    trackingParams: string
    moreContentButton?: MoreContentButton
}
// ...
export interface Content4 {
    musicTwoRowItemRenderer: MusicTwoRowItemRenderer
}

export interface MusicTwoRowItemRenderer {
    thumbnailRenderer: ThumbnailRenderer
    aspectRatio: string
    title: Title3
    subtitle: Subtitle
    navigationEndpoint: NavigationEndpoint10
    trackingParams: string
    menu: Menu2
    thumbnailOverlay?: ThumbnailOverlay
}
// ...
export interface Header3 {
    musicImmersiveHeaderRenderer: MusicImmersiveHeaderRenderer
}