// src/youtube/types/PlaylistResults_1.ts
export interface PlaylistResults_1 {
    responseContext: ResponseContext
    contents: Contents
    alerts: Alert[]
    metadata: Metadata
    trackingParams: string
    microformat: Microformat
    onResponseReceivedActions: OnResponseReceivedAction[]
    sidebar: Sidebar
}

export interface ResponseContext {
    serviceTrackingParams: ServiceTrackingParam[]
    mainAppWebResponseContext: MainAppWebResponseContext
    webResponseContextExtensionData: WebResponseContextExtensionData
}

export interface ServiceTrackingParam {
    service: string
    params: Param[]
}

export interface Param {
    key: string
    value: string
}

export interface MainAppWebResponseContext {
    datasyncId: string
    loggedOut: boolean
    trackingParam: string
}

export interface WebResponseContextExtensionData {
    hasDecorated: boolean
}

export interface Contents {
    twoColumnBrowseResultsRenderer: TwoColumnBrowseResultsRenderer
}

export interface TwoColumnBrowseResultsRenderer {
    tabs: Tab[]
}

export interface Tab {
    tabRenderer: TabRenderer
}

export interface TabRenderer {
    selected: boolean
    trackingParams: string
}

export interface Alert {
    alertWithButtonRenderer: AlertWithButtonRenderer
}

export interface AlertWithButtonRenderer {
    type: string
    text: Text
    dismissButton: DismissButton
}
// ...
export interface Metadata {
    playlistMetadataRenderer: PlaylistMetadataRenderer
}

export interface PlaylistMetadataRenderer {
    title: string
    androidAppindexingLink: string
    iosAppindexingLink: string
}

export interface Microformat {
    microformatDataRenderer: MicroformatDataRenderer
}

export interface MicroformatDataRenderer {
    urlCanonical: string
    title: string
    description: string
    thumbnail: Thumbnail
    siteName: string
    appName: string
    androidPackage: string
    iosAppStoreId: string
    iosAppArguments: string
    ogType: string
    urlApplinksWeb: string
    urlApplinksIos: string
    urlApplinksAndroid: string
    urlTwitterIos: string
    urlTwitterAndroid: string
    twitterCardType: string
    twitterSiteHandle: string
    schemaDotOrgType: string
    noindex: boolean
    unlisted: boolean
    linkAlternates: LinkAlternate[]
}

export interface Thumbnail {
    thumbnails: Thumbnail2[]
}

export interface Thumbnail2 {
    url: string
    width: number
    height: number
}

export interface LinkAlternate {
    hrefUrl: string
}

export interface OnResponseReceivedAction {
    clickTrackingParams: string
    appendContinuationItemsAction: AppendContinuationItemsAction
}

export interface AppendContinuationItemsAction {
    continuationItems: ContinuationItem[]
    targetId: string
}
// ...
export interface Sidebar {
    playlistSidebarRenderer: PlaylistSidebarRenderer
}

export interface PlaylistSidebarRenderer {
    items: Item3[]
    trackingParams: string
}